requirements.txt
black>=23.1.0
flake8>=6.0.0
pyflakes>=3.0.1

boto3==1.26.76
botocore==1.29.76

Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.9.0

PyJWT==2.6.0

psycopg2-binary==2.9.5

coverage==7.0.5

whitenoise==6.3.0

apscheduler==3.10.1